pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "android-bench"
version = "0.1.0"
description = "A repository of scripts for benchmarks LLMs against Android tasks"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "absl-py",
    "google-cloud-storage",
    "google-cloud-bigquery",
    "google-auth",
    "requests",
    "PyGithub",
    "PyYAML",
    "uv",
    "aiohttp==3.12.15",
    "datasets==4.0.0",
    "mini-swe-agent==1.11.1",
    "docker==7.1.0",
    "google-cloud-aiplatform",
    "google-genai",
    "pydantic",
    "google-cloud-batch==0.17.37",
    "matplotlib",
    "pathspec",
    "rich",
    "google-cloud-secret-manager",
    "typer<0.15.0",
]

[project.optional-dependencies]
analysis = [
    "altair",
    "black[jupyter]",
    "matplotlib",
    "numpy",
    "pandas",
    "quarto",
    "scipy",
    "seaborn",
    "statsmodels",
    "vl-convert-python",
    "shfmt-py",
]
dev = [
    "black==26.1.0",
    "pre-commit",
    "pytest",
    "pytest-xdist",
    "pytest-mock==3.15.1",
    "pytest-asyncio==1.2.0",
]

[project.scripts]
benchmark = "cli.run_benchmark:main"
build_images = "utils.docker.generate_docker_images:main"
run_task = "cli.run_task:main"
verifier = "harness.evaluation.main:run"
agent = "harness.inference.androidbench:main"
setup_oracle_agent = "utils.setup_oracle_agent:setup_oracle_agent"
setup_env = "utils.setup:main"
task_validator = "utils.task_validator.validate_task:main"
test = "pytest:main"
visualize_tasks = "utils.visualizer.commands:app"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["out*"]

[tool.pytest.ini_options]
pythonpath = ["."]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
